Rename freeform => classic
