+++ packages/block-library/src/classic/editor.scss
@@ -144,7 +144,7 @@ div[data-type="core/freeform"] .editor-block-contextual-toolbar + div {
 	padding-top: 0;
 }
 
-.freeform-toolbar {
+.block-library-classic__toolbar {
 	width: auto;
 	margin: 0 #{ -$block-padding };
 	position: sticky;
